thêm class nhỏ để xử lí đặt giá, nối phần đặt giá từ người dùng tới server lưu data

+++ auction-server/src/main/java/com/auction/server/repository/BidTransactionDaoImpl.java
@@ -78,7 +78,7 @@ public class BidTransactionDaoImpl implements IBidTransactionDAO {
         String        transactionId   = rs.getString("transaction_id");
         int           auctionId       = rs.getInt("auction_id");
         String        bidderUsername  = rs.getString("bidder_username");
-        double        bidAmount       = rs.getDouble("bid_amount");
+        long        bidAmount       = rs.getLong("bid_amount");
         java.time.LocalDateTime timestamp = rs.getTimestamp("timestamp").toLocalDateTime();
 
         return new BidTransaction(transactionId, auctionId, bidderUsername, bidAmount, timestamp);
